requirements.txt
Flask==2.1.1
python-dotenv==0.20.0
mysql-connector-python==8.0.28
boto3==1.22.0
botocore==1.25.0
PyYAML==5.4.1
awscli==1.23.0